refactor: AuthorizedLayer 디렉토리 변경
