+++ client/src/App.tsx
@@ -1,6 +1,5 @@
-import AuthorizedLayer from '@components/AuthorizedLayer';
-import UnAuthorizedLayer from '@components/UnAuthorizedLayer';
 import AccessDenied from '@pages/AccessDenied';
+import AuthorizedLayer from '@pages/AuthorizedLayer';
 import Community from '@pages/Community';
 import DM from '@pages/DM';
 import DMRoom from '@pages/DMRoom';
@@ -10,6 +9,7 @@ import NotFound from '@pages/NotFound';
 import Root from '@pages/Root';
 import SignIn from '@pages/SignIn';
 import SignUp from '@pages/SignUp';
+import UnAuthorizedLayer from '@pages/UnAuthorizedLayer';
 import UnknownError from '@pages/UnknownError';
 import React from 'react';
 import {
